Move setup into lib.rs
